ménage du layout.js
